requirements.txt
bottle==0.12.13
python-memcached==1.58
requests==2.13.0
six==1.10.0
